lua updates
