+++ include/storytime/storytime.h
@@ -1,7 +1,7 @@
 #pragma once
 
 // Entrypoint
-#include "storytime_main.h"
+#include "st_main.h"
 
 // System
 #include "system/defer.h"
@@ -36,9 +36,13 @@
 #include "script/lua_keyboard.h"
 #include "script/lua_log.h"
 #include "script/lua_mouse.h"
+#include "script/lua_serialize_deserialize.h"
 #include "script/lua_state.h"
 #include "script/lua_table.h"
 #include "script/lua_usertype.h"
+
+// System
+#include "system/json_serialize_deserialize.h"
 
 // Tiled
 #include "tiled/tiled_project.h"
