+++ src/st_main.cpp
@@ -1,4 +1,4 @@
-#include "storytime_main.h"
+#include "st_main.h"
 #include "system/service_locator.h"
 #include "system/file_reader.h"
 #include "system/event_manager.h"
