+++ src/st_pch.h
@@ -53,7 +53,6 @@
 #include "system/environment.h"
 #include "system/error.h"
 #include "system/error_signal.h"
-#include "system/json_serialize_deserialize.h"
 #include "system/log.h"
 #include "system/memory.h"
 #include "system/numbers.h"
@@ -62,7 +61,6 @@
 
 // Script
 #include "script/lua_error.h"
-#include "script/lua_push_parse.h"
 #include "script/lua_ref.h"
 #include "script/lua_utils.h"
